+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 
 ## 0.2.0 (upcoming)
 
+* Added coveralls badge
+
 ## 0.1.0 (September 2015)
 
 First release forked from Stratio Ingestion.